[ADD] create component button/accrediation
